Minimized the state in allNotebooks
