+++ src/reducers/index.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
 import userReducer from './userReducer';
-import getNotebooksReducer from './getNotebooksReducer';
+import allNotebooksReducer from './allNotebooksReducer';
 import notebooksReducer from './notebooksReducer';
 import sectionGroupsReducer from './sectionGroupsReducer';
 import notebookOrderReducer from './notebookOrderReducer'
@@ -11,7 +11,7 @@ import totalNotebookLength from './totalNotebookLengthReducer';
 export default combineReducers({
     router: routerReducer,
     users: userReducer,
-    allNotebooks: getNotebooksReducer, // used for opening notebooks
+    allNotebooks: allNotebooksReducer, // used for opening notebooks
     notebooks: notebooksReducer, // used for the open notebooks
     sectionGroups: sectionGroupsReducer,
     notebookOrder: notebookOrderReducer,
